Created a half-functioning HardAI
it doesn't work. well it works sometimes but it makes weird moves and i can't see why it isn't doing what it's supposed to. also i need to fix the uses of 'mark' and 'enemyMark' because it should change whether the AI is going first or not but i just want to fix the main issue first.